producer-air/requirements.txt
pymongo
python-dotenv
asyncio
aiohttp
requests
boto3